two checker send result to one console server test

+++ redis/redis-integration-test/src/test/java/com/ctrip/xpipe/redis/integratedtest/checker/TwoChecker2SameConsole.java
@@ -4,6 +4,7 @@ import com.ctrip.xpipe.api.endpoint.Endpoint;
 import com.ctrip.xpipe.api.pool.SimpleKeyedObjectPool;
 import com.ctrip.xpipe.api.pool.SimpleObjectPool;
 import com.ctrip.xpipe.endpoint.DefaultEndPoint;
+import com.ctrip.xpipe.endpoint.HostPort;
 import com.ctrip.xpipe.netty.commands.NettyClient;
 import com.ctrip.xpipe.pool.XpipeNettyClientKeyedObjectPool;
 import com.ctrip.xpipe.redis.core.entity.ClusterMeta;
@@ -16,16 +17,22 @@ import com.ctrip.xpipe.redis.core.protocal.cmd.PeerOfCommand;
 import com.ctrip.xpipe.redis.core.protocal.cmd.PingCommand;
 import com.ctrip.xpipe.redis.integratedtest.console.cmd.RedisStartCmd;
 import com.ctrip.xpipe.redis.integratedtest.metaserver.AbstractMetaServerMultiDcTest;
+import com.ctrip.xpipe.spring.RestTemplateFactory;
+import com.ctrip.xpipe.tuple.Pair;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 import org.springframework.context.ApplicationContext;
+import org.springframework.web.client.RestOperations;
 
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ExecutionException;
+import java.util.concurrent.TimeoutException;
+import java.util.function.BooleanSupplier;
+import java.util.function.Consumer;
 
 import static com.ctrip.xpipe.redis.checker.spring.ConsoleServerModeCondition.KEY_SERVER_MODE;
 import static com.ctrip.xpipe.redis.checker.spring.ConsoleServerModeCondition.SERVER_MODE.CONSOLE;
@@ -34,11 +41,15 @@ import static com.ctrip.xpipe.redis.console.config.impl.DefaultConsoleConfig.KEY
 public class TwoChecker2SameConsole extends AbstractMetaServerMultiDcTest {
     
     
-    ApplicationContext jqConsole;
-    ApplicationContext jqChecker;
-    ApplicationContext fraChecker;
-    RedisStartCmd jqMaster;
-    RedisStartCmd fraMaster;
+    private ApplicationContext jqConsole;
+    
+    private ApplicationContext jqChecker;
+    
+    private ApplicationContext fraChecker;
+    private RedisStartCmd jqMaster;
+    private RedisStartCmd fraMaster;
+
+    private RestOperations restOperations;
     ShardMeta getClusterMeta(String idc, String cluster, String shard) {
         return getXpipeMeta().getDcs().get(idc).getClusters().get(cluster).getShards().get(shard);
     }
@@ -56,14 +67,18 @@ public class TwoChecker2SameConsole extends AbstractMetaServerMultiDcTest {
     ZkServerMeta getZk(String idc) {
         return getXpipeMeta().getDcs().get(idc).getZkServer();
     }
+
+    private final String JQ_IDC = "jq";
     
+    private final String FRA_IDC = "fra";
+    
+    private int consolePort = 18080;
     @Before 
     public void startServers() throws Exception {
         startDb();
        
-        int consolePort = 18080;
-        final String JQ_IDC = "jq";
-        final String FRA_IDC = "fra";
+        
+        
         final String localhost = "127.0.0.1";
         String clusterName = "cluster1";
         String shardName = "shard1";
@@ -105,25 +120,48 @@ public class TwoChecker2SameConsole extends AbstractMetaServerMultiDcTest {
         extraOptions.put(KEY_CLUSTER_SHARD_FOR_MIGRATE_SYS_CHECK, "cluster-dr,cluster-dr-shard1");
         extraOptions.put(KEY_SERVER_MODE, CONSOLE.name());
         extraOptions.put("console.cluster.types", "one_way,bi_direction,ONE_WAY,BI_DIRECTION");
-        
-        startSpringConsole(consolePort, JQ_IDC, jqZk.getAddress(), Collections.singletonList("127.0.0.1:" + consolePort), consoles, metaServers, extraOptions);
+
+        jqConsole = startSpringConsole(consolePort, JQ_IDC, jqZk.getAddress(), Collections.singletonList("127.0.0.1:" + consolePort), consoles, metaServers, extraOptions);
 
         int checkerPort = 18001;
-        startSpringChecker(checkerPort++, JQ_IDC, jqZk.getAddress(), Collections.singletonList("127.0.0.1:" + consolePort), "127.0.0.2");
+        jqChecker = startSpringChecker(checkerPort++, JQ_IDC, jqZk.getAddress(), Collections.singletonList("127.0.0.1:" + consolePort), "127.0.0.2");
 
-        startSpringChecker(checkerPort++, FRA_IDC, fraZk.getAddress(), Collections.singletonList("127.0.0.1:" + consolePort), "127.0.0.3");
+        fraChecker = startSpringChecker(checkerPort++, FRA_IDC, fraZk.getAddress(), Collections.singletonList("127.0.0.1:" + consolePort), "127.0.0.3");
 
+        restOperations = RestTemplateFactory.createCommonsHttpRestTemplate(1000, 1000, 1000, 15000);
 
     }
     
+    public BooleanSupplier checkDelay(int consolePort, String srcIdc, String targetIdc) {
+        return ()-> {
+            Map<String, Map<HostPort, Object>> result = restOperations.getForObject("http://127.0.0.1:"+consolePort+"/console/cross-master/delay/bi_direction/"+srcIdc+"/cluster1/shard1", Map.class);
+            if(result == null || result.get(targetIdc) == null) return false;
+            for(Object value : result.get(targetIdc).values()) {
+                if(value instanceof  Integer) {
+                    int v = (int)value;
+                    if(v < 0 || v >= 999000) {
+                        return false;
+                    }
+                } else if(value instanceof Long) {
+                    long v = (long)value;
+                    if(v < 0 || v >= 999000) {
+                        return false;
+                    }
+                }
+            }
+            return true;
+        };
+    }
+    
     @Test
-    public void waitConsole() throws InterruptedException {
-        Thread.currentThread().join();
+    public void waitConsole() throws InterruptedException, TimeoutException {
+        waitConditionUntilTimeOut(checkDelay(consolePort, JQ_IDC, FRA_IDC), 50000, 1000);
+        waitConditionUntilTimeOut(checkDelay(consolePort, FRA_IDC, JQ_IDC), 50000, 1000);
     }
     
     @After
     public void stopServers() throws Exception {
-        jqMaster.killProcess();
-        fraMaster.killProcess();
+        if(jqMaster != null) jqMaster.killProcess();
+        if(fraMaster != null) fraMaster.killProcess();
     }
 }
